Refactored pagedAttendanceByKarkun to queries file
