+++ idreesia-common/server/graphql-api/hr/attendance/resolvers.js
@@ -17,8 +17,7 @@ import {
 } from 'meteor/idreesia-common/constants';
 import { createMonthlyAttendance } from 'meteor/idreesia-common/server/business-logic/hr/create-monthly-attendance';
 import { processAttendanceSheet } from './helpers';
-import { parse } from 'query-string';
-import { get } from 'meteor/idreesia-common/utilities/lodash';
+import { getPagedAttendanceByKarkun } from './queries';
 
 export default {
   AttendanceType: {
@@ -61,24 +60,6 @@ export default {
       return Attendances.findOne(_id);
     },
 
-    attendanceByKarkun(obj, { karkunId }, { user }) {
-      if (
-        !hasOnePermission(user._id, [
-          PermissionConstants.HR_VIEW_KARKUNS,
-          PermissionConstants.HR_MANAGE_KARKUNS,
-          PermissionConstants.HR_DELETE_KARKUNS,
-        ])
-      ) {
-        return [];
-      }
-
-      return Attendances.find(
-        {
-          karkunId,
-        },
-        { sort: { createdAt: -1 } }
-      ).fetch();
-    },
     pagedAttendanceByKarkun(obj, { queryString }, { user }) {
       if (
         !hasOnePermission(user._id, [
@@ -92,37 +73,7 @@ export default {
           totalResults: 0,
         };
       }
-      const params = parse(queryString);
-      const pipeline = [];
-      const { pageIndex = '0', pageSize = '20', karkunId } = params;
-
-      pipeline.push({
-        $match: {
-          karkunId: { $eq: karkunId },
-        },
-      });
-
-      const countingPipeline = pipeline.concat({
-        $count: 'total',
-      });
-
-      const nPageIndex = parseInt(pageIndex, 10);
-      const nPageSize = parseInt(pageSize, 10);
-      const resultsPipeline = pipeline.concat([
-        { $sort: { createdAt: -1 } },
-        { $skip: nPageIndex * nPageSize },
-        { $limit: nPageSize },
-      ]);
-
-      const karkunAttendences = Attendances.aggregate(
-        resultsPipeline
-      ).toArray();
-      const totalResults = Attendances.aggregate(countingPipeline).toArray();
-
-      return Promise.all([karkunAttendences, totalResults]).then(results => ({
-        attendance: results[0],
-        totalResults: get(results[1], ['0', 'total'], 0),
-      }));
+      return getPagedAttendanceByKarkun(queryString);
     },
     attendanceByMonth(obj, { month, categoryId, subCategoryId }, { user }) {
       if (!categoryId) return [];
